% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{profmem_prompt}
\alias{profmem_prompt}
\title{Produce a prompt-compatible string with profmem information}
\usage{
profmem_prompt(what = c("update", "suspend", "resume", "prompt", "begin",
  "end"), threshold = 10 * 1024)
}
\arguments{
\item{what}{Action to take.}

\item{threshold}{The profmem threshold to be used.}
}
\description{
Produce a prompt-compatible string with profmem information
}
\keyword{internal}
