% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprofmem-class.R
\name{total}
\alias{total}
\alias{total.Rprofmem}
\alias{subset.Rprofmem}
\title{Total number of bytes allocated}
\usage{
total(x, ...)
}
\arguments{
\item{x}{An \code{Rprofmem} object.}

\item{...}{Not used.}
}
\value{
A non-negative numeric.
}
\description{
Total number of bytes allocated
}
\keyword{internal}
