% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profmem.R
\name{profmem}
\alias{profmem}
\title{Memory profiles an R expression}
\usage{
profmem(expr, envir = parent.frame(), substitute = TRUE, ...)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envir}{The environment in which the expression should be evaluated.}

\item{substitute}{Should \code{expr} be \code{substitute()}:d or not.}

\item{...}{Not used.}
}
\value{
An object of class \code{Rprofmem}.
}
\description{
Memory profiles an R expression
}
\examples{
if (capabilities('profmem')) {

p <- profmem({
  x <- raw(1000)
  A <- matrix(rnorm(100), ncol=10)
})
print(p)

}
}

