% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRprofmem.R
\name{readRprofmem}
\alias{readRprofmem}
\title{Reads and parses an Rprofmem log file}
\usage{
readRprofmem(pathname, as = c("Rprofmem", "fixed", "raw"), drop = 0L, ...)
}
\arguments{
\item{pathname}{The Rprofmem log file to be read.}

\item{as}{Specifies in what format data should be returned.}

\item{drop}{Number of levels to drop from the top of the call stack.}

\item{...}{Not used}
}
\value{
An object of class \code{Rprofmem} (or a character vector)
}
\description{
Reads and parses an Rprofmem log file
}

